一堆修改

+++ api/src/main/java/io/dubai/modules/user/controller/ExchangeController.java
@@ -149,7 +149,7 @@ public class ExchangeController {
     @PostMapping("/sell")
     @Transactional
     public R sell(@RequestBody SellCreditsForm form, @ApiIgnore @LoginUser UserInfo userInfo) {
-        if (userInfo.getIsLockCredits().intValue() == 1 || userInfo.getIsMerchant() == 1 || userInfo.getFatherId() == 0) {
+        if (userInfo.getIsLockCredits() == 1 || userInfo.getIsMerchant() == 1 || userInfo.getFatherId() == 0) {
             return R.error(ResponseStatusEnum.CANNOT_SELL_CREDITS);
         }
         if(userInfo.getPassword() == null || "".equalsIgnoreCase(userInfo.getPassword())){
